Use 64 bit integer for kadastralegrens.lokaalID, but load it as a string, because it isn't available by OGR

+++ brk/etl/sql/create-final-tables.sql
@@ -16,7 +16,7 @@ create index bebouwing_geometrie_geom_idx on bebouwing using gist(geometrie);
 drop table bebouwing_tmp;
 
 -- KadastraleGrens
-create table kadastralegrens as select ogc_fid, gml_id, namespace, lokaalid, type, logischtijdstipontstaan::timestamptz, logischtijdstipvervallen::timestamptz, wkb_geometry as geometrie from kadastralegrens_tmp;
+create table kadastralegrens as select ogc_fid, gml_id, namespace, lokaalid::bigint, type, logischtijdstipontstaan::timestamptz, logischtijdstipvervallen::timestamptz, wkb_geometry as geometrie from kadastralegrens_tmp;
 
 alter table kadastralegrens add primary key (ogc_fid);
 alter table kadastralegrens alter column gml_id set not null;
